==> Trivariate_Example/01_analysis_roadmap.Rmd <==
---
title: "Analysis Roadmap"
author: "Haaland"
date: "03/05/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Introduction

Refactor Perry Haaland and Rose Wang's code.

Extract the data via the OpenAPI and construct a three way table. Identify any potential anomalies in the data. Complete a standard statistical analysis (TBD).


The test case we are going to work on first is for the three variables:

* `EDVisits` = Total ED/Inpatient Visits, 
* `AvgPMExposure` = Average Daily PM 2.5 Exposure (qcut), 
* `Race` = Race.

## Extracting the data through the OpenAPI

We modified the code provided by Rose in `P(XYZ) Table.Rmd` to create data frames that contain the univariate, bivariate, and trivariate tables.

**File** = `02_make_univariate_tables.Rmd`

It creates the data frames `uni_all_df`, which contains all univariate results, and `uni_df`, which contains just the three features of interest.

**File** = `03_make_bivariate_tables.Rmd`

The following data frames are created.

* `bivar_EDvsPM25_df`: `EDVisits` versus `MaxPMExposure`
* `bivar_EDvsRACE_df`: `EDVisits` versus `Race`
* `bivar_RACEvsPM25_df`: `Race` versus `MaxPMExposure`

**File**  `04_make_trivariate_table.Rmd`

It creates the data frame `tri_df`

## Validation of the tables

**File** `05_validate_tables.Rmd`

Some basic examination of marginal distributions to explore differences in results based on how the OpenAPI was programmed. 

```{r}
library(rlist)
library(httr)

httr::set_config(config(ssl_verifypeer = 0L))

get_cohort = function(cohort_def) {
  cohort_url <- "https://icees.renci.org:16339/patient/2010/cohort"
  cohort.call <- httr::POST(cohort_url,
          body=cohort_def,
          accept_json(),
          content_type("application/json"),
          encode = "json")
  cohort.text <- httr::content(cohort.call, "text")
  cohort <- jsonlite::fromJSON(cohort.text)$`return value`
  return(cohort)
}

get_cohort_id = function(cohort_def) get_cohort(cohort_def)$`cohort_id`

EDVisits = list(
  list(operator="=",value=0), 
  list(operator="=",value=1),
  list(operator="=",value=2),
  list(operator="=",value=3), 
  list(operator="=",value=4), 
  list(operator="=",value=5), 
  list(operator="=",value=6),
  list(operator="=",value=7),
  list(operator="=",value=8),
  list(operator="=",value=9),
  list(operator=">=",value=10)
)

AvgExposure = list(
  list(operator="=",value=1),
  list(operator="=",value=2),
  list(operator="=",value=3), 
  list(operator="=",value=4), 
  list(operator="=",value=5)
)

Race = list(
  list(operator="=", value="Native Hawaiian/Pacific Islander"),
  list(operator="=", value="Caucasian"), 
  list(operator="=", value="African American"), 
  list(operator="=", value="Asian"), 
  list(operator="=", value="Unknown"), 
  list(operator="=", value="American/Alaskan Native"), 
  list(operator="=", value="Other")
)

Sex2 = list(
  list(operator="=", value="Male"),
  list(operator="=", value="Female")
)

ED=list(vals=EDVisits, name="TotalEDInpatientVisits", expr=expr(TotalEDInpatientVisits))
RACE=list(vals=Race, name="Race", expr=expr(Race))
PM25=list(vals=AvgExposure, name="AvgDailyPM2.5Exposure_StudyAvg_qcut", expr=expr(`AvgDailyPM2.5Exposure_StudyAvg_qcut`))
SEX2=list(vals=Sex2, name="Sex2", expr=expr(Sex2))

cell = function(x) if (x$operator == "=") x$value else paste0(x$operator,x$value)

univariate = function(year, cohort_id, a) {
  b = SEX2
  adata = list()
  adata[[a$name]] = a$vals
  bdata = list()
  bdata[[b$name]] = b$vals
  body_bivar = toJSON(list(feature_a = adata, feature_b = bdata), auto_unbox = TRUE)
  
  bivar_call <- httr::POST(paste0("https://icees.renci.org:16339/patient/", year, "/cohort/", cohort_id, "/feature_association2"),
                     body = body_bivar,
          accept_json(),
          content_type("application/json"),
          encode = "json")
  bivar_text <- httr::content(bivar_call, "text")
  bivar_list <- jsonlite::fromJSON(bivar_text, flatten = TRUE)$`return value`
  
  fmatrix = bivar_list$columns
  univar_df = setNames(data.frame(unlist(map(a$vals, cell), use.names = FALSE), 
                        unlist(fmatrix$frequency, use.names = FALSE)), c(a$name, "Frequency"))
  return(univar_df)
}

bivariate = function(year, cohort_id, a, b) {
  adata = list()
  adata[[a$name]] = a$vals
  bdata = list()
  bdata[[b$name]] = b$vals
  body_bivar = toJSON(list(feature_a = adata, feature_b = bdata), auto_unbox = TRUE)
  
  bivar_call <- httr::POST(paste0("https://icees.renci.org:16339/patient/", year, "/cohort/", cohort_id, "/feature_association2"),
                     body = body_bivar,
          accept_json(),
          content_type("application/json"),
          encode = "json")
  bivar_text <- httr::content(bivar_call, "text")
  bivar_list <- jsonlite::fromJSON(bivar_text, flatten = TRUE)$`return value`
  
  fmatrix = bivar_list$feature_matrix
  mlist = mapply(function(bval, fmat) {
    avals = unlist(map(a$vals, cell), use.names = FALSE)
    bvalue = cell(bval)
    freq = fmat$frequency
    temp_df <- setNames(data.frame(avals, 
                        bvalue,
                        freq), c(a$name, b$name, "Frequency"))
    return(temp_df)
  }, b$vals, fmatrix, SIMPLIFY = FALSE)
  bivar_df = list.rbind(mlist)
  return(bivar_df)
}

trivariate = function(year, a, b, c) { 
  tri_cohorts <- unlist(map(a$vals, function(val) {
    tri_body_tmp = list()
    tri_body_tmp[[a$name]] = val
    tri_cohort <- get_cohort_id(toJSON(tri_body_tmp, auto_unbox = TRUE))
    return(tri_cohort)
  }), use.names = FALSE)
  tri_df <- list.rbind(mapply(function(val, tri_cohort) cbind(setNames(data.frame(valueName(val)), a$name), bivariate(year, tri_cohort, b, c)), a$vals, tri_cohorts, SIMPLIFY = FALSE))
  return(tri_df)
}
```
