==> Trivariate_Example/03_make_bivariate_tables.Rmd <==
---
title: "Make Bivariate Tables"
author: "Perry Haaland"
date: "03/05/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(jsonlite)
library(httr)
```

## Introduction

Eventually we are going to extract a three-way table and the three corresponding two-way tables using the OpenAPI, and the univariate counts for each variable separately. Because we are using the OpenAPI, we are extracting pre-binned data. The extracted values are also subject the privacy constraints of the system in terms of reporting low frequency cell counts.

The univariate counts for each variable are the univariate marginal distribution across the bins. The two-way tables are the bivariate marginal distributions. The three-way table, which is the full multivariate distribution is constrained by its marginal distributions, so we can use those to check on the correctness of the OpenAPI results.

* Before running this chunks in this script, please run the code in `02_make_univariate_tables.Rmd'.

We use the `json` format of the interface. The data set that we aer extracting from is

* Cohort 22 (all patients),
* 2010 data,
* Version 1.0.0 of ICEES.

In this file we extract the bivariate tables. The features that we are working with are as follows:

* `EDVisits` = Total ED/Inpatient Visits, 
* `MaxPMExposure` = Maximum Daily PM 2.5 Exposure (qcut), 
* `RoadwayExposure` = Major Roadway Highway Exposure.

We will extract the three bivariate combinations of the three features (Total ED/Inpatient Visits, Average Daily PM 2.5 Exposure, Major Roadway Highway Exposure) from the ICEES API. We will use the parameters of: Cohort 22 (all patients), year 2010, version 1.0.0 of ICEES data, and patient level data. We will also be using the Hypothesis Driven NxN Feature Association Functionality.

## Revisiting the relevant univariate features

My current understanding of the OpenAPI, is that the request for bivariate tables has to be hard coded specific for each feature. It seems to me that there should be a way to automate this and get all bivariate tables, or at least to get bivariate tables by just giving the feature names. That task may be left to a graduate student when we get one.

In order to do the hard coding, we need to know the bins for each feature. Here is that information. Note that there are 30 rows in the first table. After the first 10 rows, all of the counts are 10 or less. It seems like the higher ones should be binned before proceeding to the two-way tables.


## Creating the bivariate tables

In order to make the code clearer, and perhaps pave the way for later automation, we create some variables that will be used in the OpenAPI calls. The first call will be `TotalEDInpatientVisits` versus `MaxDailyPM2.5Exposure_qcut`.

* Note: The range of TotalEDInpatientVisits, as we saw above goes up to 51 visit. Because all of the values of 10 or more are relatively small, we don't expect them to be returned as actual numbers but will be replaced by NA's because of the privacy constraints. So in this example we are just going to ignore the upper range. 

* An alternative to this strategy would be to group all of the 10 visits or above subjects into one group. We leave that as an exercise for the future.

### Table 1. Total ED Inpatient Visits versuys Maximum Daily PM 2.5 Exposure
I picked a generic name for the table, for the time being. There probably should be more internal error checking.

```{r}
cohort.id <- get_cohort_id('{}')
```

```{r, message=FALSE}
bivar_EDvsPM25_df = bivariate(2010,cohort.id,ED,PM25)
bivar_EDvsPM25_df
```

### Table 2. Total ED Inpatient Visits verus Major Roadway Highway Exposure

The second call will be `TotalEDInpatientVisits` versus `MajorRoadwayHighwayExposure`.

```{r, message=FALSE}
bivar_EDvsRACE_df = bivariate(2010, cohort.id, ED, RACE)
bivar_EDvsRACE_df
```

### Table 3. Major Roadway Highway Exposure versus Maximum Daily PM2.5 Exposure

The third call will be for `MajorRoadwayHighwayExposure` vs `MaxDailyPM2.5Exposure_qcut`

```{r, message=FALSE}
bivar_RACEvsPM25_df = bivariate(2010, cohort.id, RACE, PM25)
bivar_RACEvsPM25_df
```

## Save the resulting data frames


```{r}
save(bivar_EDvsPM25_df, file = "RData/bivar_EDvsPM25_df.RData")
save(bivar_EDvsRACE_df, file = "RData/bivar_EDvsRACE_df.RData")
save(bivar_RACEvsPM25_df, file = "RData/bivar_RACEvsPM25_df.RData")
```
