==> Trivariate_Example/04_make_trivariate_table.Rmd <==
---
title: "Make Trivariate Table"
author: "Perry Haaland"
date: "03/05/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(jsonlite)
library(httr)
```

## Trivariate Counts

In this file we make three-way, or trivariate table. In previous files we created the univariate table and the three bivariate tables for Roadway Exposure, Emergency Room Visits, and PM2.5 exposure. In order create the three-way table, we narrow the cohort of each of our calls to the API to one bin of the Total ED/Inpatient Visits feature. Thus, we first have to create each of the cohorts that we will be using. 



Next, we can use the cohorts we just created to pull the bivariate distribution between Major Roadway Highway Exposure and Max Daily PM2.5 Exposure. We will use the parameters of: varying cohorts, year 2010, version 1.0.0 of ICEES data, and patient level data. We will also be using the Hypothesis Driven NxN Feature Association Functionality.

```{r trivariate table, message=FALSE}
tri_df = trivariate(2010, ED, RACE, PM25)
```

The table looks as follows:

```{r}
tri_df
```

The total number of subject reported in the trivariate table is as follows:

```{r}
sum(tri_df$Frequency)
```

## Save the resulting data frame


```{r}
save(tri_df, file = "RData/tri_df.RData")
```
