==> Trivariate_Example/05_validate_tables.Rmd <==
---
title: "Validation of Tables"
author: "Haaland"
date: "03/11/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)

```


## Load the tables

In files 2-4, we created and saved the univariate, bivariate and trivariate tables for the variables:

* `EDVisits` = Total ED/Inpatient Visits, 
* `MaxPMExposure` = Maximum Daily PM 2.5 Exposure (qcut), 
* `RoadwayExposure` = Major Roadway Highway Exposure.

The three pairwise combinations of binary tables are labeled as follows:

* `bivar_EDvsME_df`: `EDVisits` versus `MaxPMExposure`
* `bivar_EDvsRE_df`: `EDVisits` versus `RoadwayExposure`
* `bivar_REvsME_df`: `RoadwayExposure` versus `MaxPMExposure`

We start by loading the data.

```{r}
load("RData/univar_ED_df.RData")
load("RData/univar_RACE_df.RData")
load("RData/univar_PM25_df.RData")
load("RData/bivar_EDvsPM25_df.RData")
load("RData/bivar_EDvsRACE_df.RData")
load("RData/bivar_RACEvsPM25_df.RData")
load("RData/tri_df.RData")
load("RData/uni_df.RData")
```



## Brief examination of bivariate tables

## Compare bivariate marginals to univariate

To verify the correctness of the data, we can compare the marginal distributions from the bivariate tables to the univariate distributions.

### EDVisits

First we consider the univariate and marginal distributions for EDVisits. We calculate the marginal distributions from the bivariate and trivariate tables and compare to the univariate values. The results are be the same.

```{r}
addFeature = function(df, a) {
  cbind(data.frame(Feature=a$name), rename(df, Value=a$name))
}

compare = function(a, bva, bvb, uv) {
  bv1_df = bva %>%
    group_by_at(a$name) %>%
    summarize(bv1Frequency = sum(Frequency)) %>%
    select(Value = a$name, bv1Frequency)
  bv2_df = bvb %>%
    group_by_at(a$name) %>%
    summarize(bv2Frequency = sum(Frequency)) %>%
    select(Value = a$name, bv2Frequency)
  tr1_df = tri_df %>%
    group_by_at(a$name) %>%
    summarize(triFrequency = sum(Frequency)) %>%
    select(Value = a$name, triFrequency)
  un_df = addFeature(uv, a) %>%
    rename(uniFrequency = Frequency)
  comp1_df = merge(merge(merge(un_df,bv1_df, by="Value"),bv2_df, by="Value"),tr1_df, by="Value") %>%
    select(Feature, Value, uniFrequency, bv1Frequency, bv2Frequency, triFrequency)
  return(comp1_df)
}
```

```{r}
compare(ED, bivar_EDvsPM25_df, bivar_EDvsRACE_df, univar_ED_df)
```

We can see by inspection, that the results look correct. We test that formally below:

```{r}
comp1_df %>%
  transmute(Feature = Feature, Value = Value, agree = ((uniFrequency == bv1Frequency) & 
                                                         (uniFrequency == bv2Frequency) &
                                                        (uniFrequency == triFrequency)))
```

### RoadwayExposure

First we consider the univariate and marginal distributions for RoadwayExposure. We calculate the marginal distributions from the bivariate and trivariate tables and compare to the univariate values. In this comparison, we see differences among the data extraction approaches. I show below that this is related to the truncation of the ED Visits variable.

```{r}
comp2_df = compare(RACE, bivar_EDvsRACE_df, bivar_RACEvsPM25_df, univar_RACE_df)
comp2_df
```

The univariate frequency and the first bivariate frequency give different values as can be seen above. The difference in the total is

```{r}
sum(comp2_df$uniFrequency) - sum(comp2_df$bv1Frequency)
```

If we look at the number of truncated values based on ED visits, we get the same number:

```{r}
uni_df %>%
  filter(Feature == ED$name,
         as.numeric(Value) > 9) %>%
  summarize(Truncated_Number = sum(Frequency))

```

The univariate frequency and the trivariate marginal frequency give different values as can be seen above. The difference in the total is

```{r}
sum(comp2_df$uniFrequency) - sum(comp2_df$triFrequency)
```

This difference can also be seen to be the same as for the bivariate table. So we conclude that the differences in marginals are due to the truncation of ED Visits.

### MaxPMExposure

First we consider the univariate and marginal distributions for MaxPMExposure. We calculate the marginal distributions from the bivariate and trivariate tables and compare to the univariate values. In this comparison, we see same pattern of differences in the marginal frequences as above.

```{r}
comp3_df = compare(PM25, bivar_EDvsPM25_df, bivar_RACEvsPM25_df, univar_PM25_df)
comp3_df
```

```{r}
sum(comp3_df$uniFrequency) - sum(comp3_df$bv1Frequency)
```
